+++ src/DumpIO/SetPos.cpp
@@ -12,18 +12,18 @@ void DumpIO::SetPos(std::vector<std::vector<double> >& pos_t_) {
   set_pos = true;
 
   if (set_id) {
-    if (n_p != pos_t_.size()) {
+    if (size_t(n_p) != pos_t_.size()) {
       std::cout << "Number of particles mismatch!";
-      assert(n_p == pos_t_.size());
+      assert(size_t(n_p) == pos_t_.size());
     }
   } else {
     n_p = pos_t_.size();
   }
 
   if (set_bb) {
-    if (d != pos_t_[0].size()) {
+    if (size_t(d) != pos_t_[0].size()) {
       std::cout << "Dimensions mismatch!";
-      assert(d == pos_t_[0].size());
+      assert(size_t(d) == pos_t_[0].size());
     }
   } else {
     d = pos_t_[0].size();
